refactor class into mixins and separate modules
